% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{letter}
\alias{letter}
\alias{report}
\title{Output format for USYD letter}
\usage{
letter(...)

report(...)
}
\arguments{
\item{\dots}{Arguments passed to \code{\link[bookdown]{pdf_document2}}.}
}
\value{
An R Markdown output format object.
}
\description{
Each function is a wrapper for \code{\link[bookdown]{pdf_document2}} to
produce documents in USYD style.
}
\author{
Rob J Hyndman (modified to USYD by Emi Tanaka)
}
